# Conflicts:
#	resources/views/surveyView.blade.php

+++ resources/views/surveyView.blade.php
@@ -27,6 +27,11 @@ $userCanEditDueToRole
             right: 4%;
         }
 
+        #dropdown-menu_name2 {
+            position: relative;
+            left: -0.5%;
+        }
+
         #dropdown_club {
             position: absolute;
             overflow: visible;
@@ -110,10 +115,7 @@ $userCanEditDueToRole
         <div class="panel-title-box">
             <h4 class="panel-title">
                 {{ $survey->title }}
-                @if($userId == $survey->creator_id OR
-                    $userGroup == "admin" OR
-                    $userGroup == "marketing" OR
-                    $userGroup == "clubleitung")
+                @if($userId == $survey->creator_id OR $userCanEditDueToRole)
                     <a href="{{$survey->id}}/edit"
                        style="float: right"
                        class="btn btn-default btn-sm"
@@ -140,7 +142,8 @@ $userCanEditDueToRole
     <br>
     <br>
 
-    {!! Form::open(['action' => ['SurveyAnswerController@store', $survey->id]]) !!}
+
+     {!! Form::open(['action' => ['SurveyAnswerController@store', $survey->id], 'class' => 'store', 'name' => 'store']) !!}
 
     <div class="panel panel-warning">
         @if( $survey->password != '')
@@ -157,6 +160,13 @@ $userCanEditDueToRole
         <div class="nameToQuestion">
             <div class="panel" id="panelNoShadow">
                 <div id="survey-answer" class="table-responsive-custom">
+
+                    <input hidden id="get_row" value="">
+                    <input type="hidden" id="hdnSession_userName" value="{{Session::get('userName')}}">
+                    <input type="hidden" id="hdnSession_userClub" value="{{Session::get('userClub')}}">
+                    <input type="hidden" id="hdnSession_userID" value="{{Session::get('userId')}}">
+                    <input type="hidden" id="hdnSession_oldContent" name="hidden_oldContent[]" value="">
+
                     <table class="table table-striped table-bordered table-condensed table-responsive-custom">
                         <thead>
                         <tr>
@@ -195,13 +205,15 @@ $userCanEditDueToRole
                                 {{--autocomplete for clubs is not working right now--}}
                                 <div id="dropdown_club" class="dropdown">
                                     <div class="btn-group col-md-8 no-padding">
-                                        {!! Form::text('club', null, ['class' => 'form-control', 'id' => 'club', 'placeholder' => 'mein Club', 'autocomplete' => 'off', 'required' => true, 'oninvalid' => 'setCustomValidity(\'Bist Du mitglied in einem Club?\')', 'oninput' => 'setCustomValidity(\'\')']) !!}
+                                        {!! Form::text('club', null, ['class' => 'form-control', 'id' => 'club', 'placeholder' => 'mein Club', 'autocomplete' => 'off', 'oninvalid' => 'setCustomValidity(\'Bist Du mitglied in einem Club?\')', 'oninput' => 'setCustomValidity(\'\')']) !!}
                                     </div>
                                     <ul id="dropdown-menu_club" class="dropdown-menu dropdown-club"></ul>
                                 </div>
                             </td>
                             @foreach($questions as $key => $question)
-                                <td class="question" style="vertical-align: middle;">
+                                <input type="hidden" id="field_type{{$question->order-1}}" value="{{$question->field_type}}" />
+                                <input type="hidden" id="question_order" value="{{$question->order}}" />
+                                <td class="question{{$question->order}}" style="vertical-align: middle;">
                                     @if($question->field_type == 1)
                                         <!-- Freitext -->
                                         @if(!$question->is_required)
@@ -213,46 +225,46 @@ $userCanEditDueToRole
                                         @endif
                                     @elseif($question->field_type == 2)
                                         <!-- Ja/Nein -->
-                                        {{ Form::radio('answers['.$key.']', 1) }} Ja
+                                        {{ Form::radio('answers['.$key.']', 1, '' , ['id' => 'radio'.$question->order.'-0']) }} Ja
                                         @if(!$question->is_required)
                                             <!--Answer not required-->
-                                            {{ Form::radio('answers['.$key.']', 0) }} Nein
-                                            {{ Form::radio('answers['.$key.']', -1, true)}} keine Angabe
+                                            {{ Form::radio('answers['.$key.']', 0, '' , ['id' => 'radio'.$question->order.'-1']) }} Nein
+                                            {{ Form::radio('answers['.$key.']', -1, true, ['id' => 'radio'.$question->order.'-2'])}} keine Angabe
                                         @else
                                             <!--Answer* required-->
-                                            {{ Form::radio('answers['.$key.']', 0, true) }} Nein
+                                            {{ Form::radio('answers['.$key.']', 0, true, ['id' => 'radio'.$question->order.'-0']) }} Nein
                                         @endif
                                     @elseif($question->field_type == 3)
                                         <!-- Dropdown -->
-                                        <select class="form-control" name="answers[{{$key}}]" style="font-size: 13px;">
+                                        <select class="form-control" id="options{{$question->order}}" name="answers[{{$key}}]" style="font-size: 13px;">
                                             @if(!$question->is_required)
                                                 <option>keine Angabe</option>
                                             @endif
                                             @foreach($question->getAnswerOptions as $answerOption)
-                                                <option>{{$answerOption->answer_option}}</option>
+                                                <option id="option">{{$answerOption->answer_option}}</option>
                                             @endforeach
                                         </select>
                                     @endif
                                 </td>
                             @endforeach
                             <td class="tdButtons " id="panelNoShadow">
-                                {{--{!! Form::submit('<i class="fa fa-pencil"></i>', ['type' => 'submit', 'class' => 'btn btn-primary btn-margin', 'style' => 'display: inline-block;']) !!}
+                                {{--{!! Form::submit('<i class="fa fa-pencil"></i>', ['type' => 'submit', 'class' => 'btn btn-primary btn-margin', 'style' => 'display: inline-block;', 'fo]) !!}
                                 --}}
-                                <button type="submit" class="btn btn-primary btn-margin" id="noMarginMobile"
-                                        style="display: inline-block;"><i class="fa fa-floppy-o"></i></button>
-                                {!! Form::close() !!}
+                                <input type="submit" class="btn btn-primary btn-margin fa fa-floppy-o" id="noMarginMobile" value=""
+                                        style="display: inline-block; height: 34px;">
+                               {!! Form::close() !!}
                             </td>
                         </tr>
-
+                        {!! Form::open(['action' => ['SurveyAnswerController@update', $survey->id,  'id' => '' ], 'class' => 'update']) !!}
                         @if(!$survey->is_anonymous OR $userId == $survey->creator_id)
                             @if(!$survey->show_results_after_voting OR $userParticipatedAlready)
                                 @foreach($answers as $key => $answer)
-                                    <tr class="row{{$answer->id}}">
-                                        <td>
+                                    <tr class="row{{$answer->id}}" id="{{$answer->id}}">
+                                        <td class="singleAnswer">
                                             @include('partials.surveyAnswerStatus')
                                             {{$answer->name}}
                                         </td>
-                                        <td>
+                                        <td class="singleAnswer">
                                             @if(!empty($answer->club))
                                                 {{$answer->club}}
                                             @else
@@ -265,26 +277,32 @@ $userCanEditDueToRole
                                             </td>
                                         @endforeach
                                         @if($userId == $answer->creator_id OR $userCanEditDueToRole OR empty($answer->creator_id))
-                                        <!--Edit Delete Buttons-->
-                                            <td class="tdButtons panel" id="panelNoShadow">
-                                                {{--not working at the moment--}}
-                                                {{--<a href="#"--}}
-                                                   {{--class="editButton btn btn-primary "--}}
-                                                   {{--data-toggle="tooltip"--}}
-                                                   {{--data-placement="bottom">--}}
-                                                    {{--<i class="fa fa-pencil" id="{{$answer->id}}"></i>--}}
-                                                {{--</a>--}}
-                                                <a href="{{$survey->id}}/answer/{{$answer->id}}"
-                                                   class="btn btn-default deleteRow"
-                                                   data-toggle="tooltip"
-                                                   data-placement="bottom"
-                                                   data-method="delete"
-                                                   data-token="{{csrf_token()}}"
-                                                   rel="nofollow"
-                                                   data-confirm="Möchtest Du diese Antwort wirklich löschen?">
-                                                    <i class="fa fa-trash"></i>
-                                                </a>
-                                            </td>
+                                            @if($survey->deadline >= date("Y-m-d H:i:s") OR $userCanEditDueToRole)
+                                            <!--Edid Delete Buttons-->
+                                                <td class="tdButtons panel" id="panelNoShadow">
+                                                    <input href="#"
+                                                       class="editButton btn btn-primary fa fa-pencil"
+                                                       id="editButton{{$answer->id}}"
+                                                       value=""
+                                                       style="height: 34px; width: 43px;"
+                                                       type="button"
+                                                       data-toggle="tooltip"
+                                                       data-placement="bottom"
+                                                       onclick="change_to_submit({{$answer->id}}); get_answer_row({{$answer->id}});">
+                                                    <i id="spinner{{$answer->id}}" class="fa fa-spinner fa-spin fa-2x hidden" style="position: absolute; left: 8px; bottom: 3px;"></i>
+
+                                                    <a href="{{$survey->id}}/answer/{{$answer->id}}"
+                                                       class="btn btn-default deleteRow"
+                                                       data-toggle="tooltip"
+                                                       data-placement="bottom"
+                                                       data-method="delete"
+                                                       data-token="{{csrf_token()}}"
+                                                       rel="nofollow"
+                                                       data-confirm="Möchtest Du diese Antwort wirklich löschen?">
+                                                        <i class="fa fa-trash"></i>
+                                                    </a>
+                                                </td>
+                                            @endif
                                         @else
                                             <td class="tdButtons panel" id="panelNoShadow">
                                             </td>
@@ -292,6 +310,7 @@ $userCanEditDueToRole
                                     </tr>
                                 @endforeach
 
+                                {{ Form::close() }}
                                 <!-- Start of evaluation -->
                                 {{-- shows a statistic of answers of the users who already took part in the survey--}}
 
@@ -300,10 +319,10 @@ $userCanEditDueToRole
                                 @foreach($answers as $key => $answer)
                                     @if($i == 0)
                                         <tr>
-                                            <td class="transparent background">bg</td>
-                                            <td class="transparent background">bg</td>
+                                            <td class="transparent background">&nbsp;</td>
+                                            <td class="transparent background">&nbsp;</td>
                                             @foreach($answer->getAnswerCells as $cell)
-                                                <td class="transparent background">bg</td>
+                                                <td class="transparent background">&nbsp;</td>
                                             @endforeach
                                         </tr>
                                         <tr id="evaluation">
@@ -428,15 +447,5 @@ $userCanEditDueToRole
             </div>
         @endif
     @endif
-
-
-    <script>
-        $(document).ready(function () {
-            $('#surveyAnswerForm').formValidation();
-        });
-        $(document).ready(function () {
-            $('#surveyAnswerFormMobile').formValidation();
-        });
-    </script>
 
 @stop
